break the lib into some side modules.
